fix(ui): barre supérieure et paramètres — chevauchement, badge d'année, Démarrage, toggle, date

+++ src/SamaEcole.Web/wwwroot/js/ui-components.js
@@ -498,7 +498,9 @@ document.addEventListener('alpine:init', () => {
             const dateStr = now.toLocaleDateString('fr-FR', {
                 weekday: 'long',
                 year: 'numeric',
-                month: 'long',
+                // 'short' : « sept. », « janv. », « févr. »… — n'abrège QUE les mois longs (mai, juin,
+                // août, mars restent entiers), ce qui resserre le pill sans rien tronquer d'inutile.
+                month: 'short',
                 day: 'numeric'
             });
             const timeStr = now.toLocaleTimeString('fr-FR', {
